OCRmypdf_plugins/OCRmypdf_plugins/ocrmypdf-doctr-master/pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocrmypdf-doctr"
dynamic = ["version"]
description = "docTR (Document Text Recognition) plugin for OCRmyPDF"
requires-python = ">=3.10"
license = {text = "MPL-2.0"}
keywords = ["ocr", "pdf", "doctr", "ocrmypdf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Capture :: Scanners",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "ocrmypdf>=14.0.0",
    "python-doctr>=1.0.0",
    "pillow>=9.0.0",
    "torch>=2.0.0",
]

[project.entry-points.ocrmypdf]
doctr = "ocrmypdf_doctr.plugin"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/ocrmypdf_doctr/_version.py"
